Refactor connection manager logic.
