+++ src/server/browser/server_browser.ts
@@ -27,7 +27,7 @@ import {
   BrowserMessageWriter,
 } from 'vscode-languageserver/browser';
 import {initServer} from '../init';
-import {ConnectionManager} from '../../common/connection_manager';
+import {CommonConnectionManager} from '../../common/connection_manager';
 import {WebConnectionFactory} from '../connections/browser/connection_factory';
 import {MessageHandler} from '../../worker/message_handler';
 
@@ -35,7 +35,7 @@ const messageReader = new BrowserMessageReader(self as unknown as Worker);
 const messageWriter = new BrowserMessageWriter(self as unknown as Worker);
 
 const connection = createConnection(messageReader, messageWriter);
-const connectionManager = new ConnectionManager(
+const connectionManager = new CommonConnectionManager(
   new WebConnectionFactory(connection)
 );
