+++ src/worker/compile_query.ts
@@ -5,7 +5,7 @@
  *  LICENSE file in the root directory of this source tree.
  */
 
-import {ConnectionManager} from '../common/connection_manager';
+import {ConnectionManager} from '../common/types/connection_manager_types';
 import {CellData, FileHandler} from '../common/types/file_handler';
 import {createModelMaterializer} from './create_runnable';
 import {DocumentMetadata} from '../common/types/query_spec';
